refactor: move utils/ => helpers.ts
